+++ tests/api/test_generate.py
@@ -15,6 +15,9 @@ EXTERNAL_CONFIG_BODY = {
         "aws_access_key_id": getenv("AWS_ACCESS_KEY_ID", ""),
         "aws_secret_access_key": getenv("AWS_SECRET_ACCESS_KEY", ""),
     },
+    "db": {
+        "connection_string": "postgresql+psycopg2://postgres:postgres@postgres-test:5432/postgres_example?"
+    },
     "okta": {
         "orgUrl": "https://dev-78908748.okta.com",
         "token": getenv("OKTA_CLIENT_TOKEN", ""),
@@ -24,7 +27,8 @@ EXTERNAL_CONFIG_BODY = {
 
 @pytest.mark.external
 @pytest.mark.parametrize(
-    "generate_type, generate_target", [("systems", "aws"), ("systems", "okta")]
+    "generate_type, generate_target",
+    [("systems", "aws"), ("systems", "okta"), ("datasets", "db")],
 )
 def test_generate(
     test_config: FidesctlConfig,